Deliver only one initial message to new ObservableQuery subscribers.
PR #6426 shows why delivering this.lastResult before this.lastError was
problematic (when this.lastError is set), but I think there's a slightly
better way to solve the problem than reversing the order of the delivery.

When an ObservableQuery has received an error, that error should be
delivered to new subscribers instead of (not just before) the most recent
result, because delivering an error immediately followed by a result is
either useless (if the subscriber immediately unsubscribes due to the
error) or misleading (if the immediate delivery of this.lastResult hides
this.lastError, despite the error being more recent than the result).

Note that updateLastResult clears this.lastError when a new result is
received with no errors, so the ObservableQuery will go back to delivering
normal results to new subscribers as soon as it receives another result.