Revert "Avoid network after incomplete optimistic cache results. (#6419)"
This reverts commit 94509386e984e093bd67f2ca995fcd4d89574b95.

The logic of diff.optimistic needs rethinking. Any query could read from
the cache at a time when optimistic updates are in progress, and that
possibility should not affect the network behavior of the query.

I had hoped that the dependency tracking system would serve to protect
queries whose field dependencies were unrelated to the optimistic updates,
but I have now observed at least one case where this hope was badly
mistaken, while debugging an endless loading spinner issue with the latest
versions of @apollo/client (rc3-rc.8) in the studio.apollographql.com
application.

As I mentioned in #6419, I was not able to reproduce the original scenario
using an ordinary optimistic mutation, so I think we should wait until
someone can provide a more realistic reproduction of the problem:
https://github.com/apollographql/apollo-client/pull/6419#discussion_r437709911

I have left the test in place, disabled with itAsync.skip, so that we can
revisit this functionality in the future.