Document undefined `data` cases for mutations
Unfortunately `data` legitimately can be undefined in some edge cases
for mutations. It is sligthly annoying because in most normal cases it
will never be undefined, but this cannot be reflected in typing as we
have to also support those edge cases.

So end-user will have to use `result.data!.foo.bar` syntax in most cases
if they enable strict null checks.

Fixes #5662