Avoid defaulting to empty data in relayStylePagination read function.
Fixes #6611, thanks to this astute comment from @fracmak:
https://github.com/apollographql/apollo-client/issues/6611#issuecomment-662676848

Returning undefined from a field read function indicates to the
StoreReader that the field is missing, and usually means the query will be
fetched from the server. When we haven't yet written any data to a field
paginated using relayStylePagination, requesting data from the server is
almost always better than providing any kind of default information, so we
should avoid using makeEmptyData() as a default value.