Update why-apollo.mdx
I may be confused about how to name queries, but if the Apollo Cache stores data based on `id` and `__typename`, then the actual name of the query would not need to be the same as the plural form of the query. And in this case, I think the more semantically appropriate name would be GET_DOG. 

If I am misunderstanding naming rules, then it may be helpful to clarify the rules.