Update cache-field-behavior.md
The official documentation lacked an example and description of how to
handle multiple parameters within the read() function when defining
typePolicies. The proposed change addresses that by both showcasing that
it is possible to handle multiple parameters and how to do so (using a
syntax that was previously undocumented).

I struggled for a while when it came to understanding how to handle this
use case, so I am more than happy to share my findings while improving
the docs with the hope of helping others that might find themselves in
the same situation. 😃