Fix optimistic mutation tests, again.
This essentially undoes commit b2391243df0bbe91408dd5f3acf1c06276ba731e,
which was introduced as part of PR #6221. I remember thinking, as I made
that commit, "This change shouldn't be necessary, but it seems harmless
enough." As it turns out, changing these tests was not necessary, but was
instead a symptom of overreacting to the delivery of optimistic mutation
results. I'm happy to be putting things back the way they were.