Add subscriptions-transport-ws in get-started.mdx
As `@apollo/client` has required peer dependency on `subscriptions-transport-ws@^0.9.0`, it should be installed as a peer alongside `graphql` on any new project. Otherwise, the following warning will appear upon installing:

```
npm WARN @apollo/client@3.1.3 requires a peer of subscriptions-transport-ws@^0.9.0 but none is installed. You must install peer dependencies yourself.
```

Side note: Ideally, `subscriptions-transport-ws` would be an optional peer dependency as it is likely to be used only in a minority of projects.