Make keyArgs tolerant of optional arguments.
The keyArgs:[...] configuration for field policies should be able to
include arguments that are optional, while still specifying an ordering of
all possible key arguments for serialization purposes. When an optional
argument is not provided, it will simply not be included in the serialized
storeKeyName suffix, and no exception will be thrown.

Should address #6973.