Avoid all sub-package imports from graphql npm package. (#7185)
Since we don't control the contents of the graphql package, and
nested subdirectories within the graphql package do not have their
own package.json files with "main" and "module" entry points (like
our @apollo/client/* sub-packages do), it is unfortunately not safe
to reach into the package to import specific items, as demonstrated
by #7184.

This change could cause a bundle size regression for bundlers that
do not perform any tree-shaking, but it's the responsibility of the graphql
package maintainers to support selective sub-package imports, if they
want to. Until that happens, @apollo/client must import from the graphql
package in the only supported way: directly from the top-level package.