Update apollo-cache-persist to apollo3-cache-persist. (#7174)
The apollo-cache-persist package has been updated to work with Apollo
Client 3, but it has a new package name (not unlike the rename from
apollo-client to @apollo/client). I've seen multiple folks installing the
older apollo-cache-persist with @apollo/client and noticing problems, so I
want to make sure the docs recommend the new package.