Simplify the useLazyQuery example of the Queries section
This PR simplifies the useLazyQuery example of the Queries section as
the same as the useLazyQuey example of the Hook section.
People may think they should do the same when they use useLazyQuery
but I don't see any benefit in it, so I think the extra state
for a dog should be removed.