Reactivate reactive variables when InMemoryCache gets new watchers.
Fixes a bug introduced by #7279, which was triggered by calling
this.cancel() before re-watching in the QueryInfo#updateWatch method,
causing cache.watches.size to drop to zero temporarily, thereby invoking
forgetCache(cache). When this happened, reactive variables would stop
broadcasting updates to the forgotten cache, even though the unwatching
was only momentary, and would not begin broadcasting again until the
association was reestablished by other means.

This new implementation uses a WeakMap to remember the association between
caches and sets of reactive variables, which makes it possible to recall
those associations later, provided the cache has not been garbage
collected in the meantime.