Attempt to remove ink dependency from docs/package.json.
Closes #7473, if nothing breaks.

This dependency was added in bf9d073fec8512225aec6e03427348260783c325 as
part of #6429. Back then, `gatsby` and `gatsby-cli` depended on the `ink`
package, but I can't find any current usage of the `ink` package outside
of the devDependencies of those Gatsby packages, so I believe it should be
safe to remove it.