Make refetch replacement behavior opt-in without adding refresh.
This reverts commit d87333bff032ec1fab703e2cb28134200d82e340 and
introduces a new option for WatchQueryOptions called refetchPolicy.

Because of the backwards compatibility concerns raised in #7810, the
default setting for refetchPolicy needs to be "merge", though we hope
developers will switch to refetchPolicy: "overwrite" using the explicit
option once Apollo Client v3.4 is released, to achieve the behavior I
described in my comment https://github.com/apollographql/apollo-client/issues/7491#issuecomment-772115227.

We try to avoid introducing ad hoc new options, especially when we think
one of the behaviors is preferable (overwriting, in this case). However,
since not all observableQuery.refetch call sites are under the developer's
control (for example, see refetchQueries after a mutation), it seemed
important to have a way to _alter_ the behavior of refetch, rather than a
new alternative method (refresh).

If one of these behaviors ("merge" or "overwrite") becomes a clear
favorite after Apollo Client v3.4 is released, we may either swap the
default or remove the refetchPolicy option altogether in Apollo Client 4.