Broadcast watches before transaction when onDirty provided.
When an options.onDirty callback is provided to cache.batch (#7819), we
want to call it with only the Cache.WatchOptions objects that were
directly affected by options.transaction, so it's important to broadcast
watches before the transaction, to flush out any pending watches waiting
to be broadcast. See provided tests for an example where this matters.