Handle already-dirty watches more gracefully in cache.batch.
PR #7904 was an important change to make sure the cache.batch onDirty
function gets called for all/only those watchers actually affected by the
batch operation, and not for any other cache watchers that happen to be
waiting to receive a broadcast.

However, in my initial implementation of this functionality, unrelated
watchers would receive pending broadcasts immediately before the
cache.batch transaction, which might be surprising because those unrelated
broadcasts appear to be triggered by calling cache.batch, but are not
included in the onDirty set.

Instead, I think we can make #7904 "silent" in the sense that the
unrelated cache watchers are properly excluded from the onDirty set (since
we call broadcastWatches before the transaction, as before), but they do
not immediately receive a broadcast during cache.batch.

To make this work, the unrelated watchers need to end up having the same
dirtiness status after the cache.batch transaction as before the
transaction, so those watchers behave as if cache.batch never happened,
and will be swept up by the _next_ broadcastWatches.