Make "active" exclude cache-only queries only for reFetchObservableQueries.
Queries with `fetchPolicy: cache-only` are certainly not "inactive" (you
just might not want to refetch them from the network), so excluding them
is a choice for the reFetchObservableQueries method to make.

Note: we are currently planning to deprecate reFetchObservableQueries
and remove it in Apollo Client v4, so this logic will disappear from the
codebase after that.

The new client.refetchQueries API provides a much more flexible way to
specify which queries you want to include, including the ability to
filter/skip them in arbitrary ways in the onQueryUpdated function.