Return use{Loading,NetworkStatus,Error,Data} hooks from useObservableQuery.
This design lets the component selectively depend on ObservableQuery
metadata that it cares about, rather than always rerendering the
component whenever any item of that metadata has been updated.

For example, the useLoading hook uses useReactiveVar to read from
result.loadingVar, which is kept in sync with the result.loading
properties of any results received by the ObservableQuery. Likewise for
useNetworkStatus, useError, and useData, though useData is especially
discouraged because it defeats the main purpose of useObservableQuery,
which is to avoid rerendering the component tree every time anything
triggers a change in result.data.

The ObservableQuery itself is now returned as the result.observable
property from useObservableQuery, which gives us room to add additional
sibling properties in the future.