Use setTimeout instead of microtask to delay Concast cancellation.
Inspired by @cornedor's comment:
https://github.com/apollographql/apollo-client/issues/7608#issuecomment-882418482

Should help with some instances of issue #7608, specifically those with
global unhandled promise rejection errors. I don't know if this is the
full solution, but I believe it is helpful and safe to introduce a delay
before unsubscribing, giving other subscribers a chance to subscribe,
thereby saving the Concast from needing to be cancelled.