To avoid errors like "Cannot add property key, object is not extensible"
In apollo client 3.0.2 and up if you want to do data.someCacheFieldpush() where data is the result of a readQuery call. It gives an error like mentioned in the title. I think it happens due to apollo cache objects being immutable directly.  I was able to solve the issue by deep copying the data object and its associated fields. Skipping push and using spread operator does not cause the issue at all.