Update react-native.md
Whilst i was followint this example, I consoled log the data and the error:   const {data, loading, error} = useQuery(USERS_QUERY);
And the error was for network request failing. 
After passing my ip address in the uri, it worked perfectly so I am changing this docs for better clarity and performance whilst using apollo client and server as the backend.