Add direct dependency on acorn@8.6.0 to fix Netlify npm failures.
The specific failure is an integrity hash mismatch:

  5:44:05 PM: npm ERR! code EINTEGRITY
  5:44:05 PM: npm ERR! sha512-yXbYeFy+jUuYd3/CDcg2NkIYE991XYX/bje7LmjJigUciaeO1JR4XxXgCIV1/Zc/dRuFEyw1L0pbA+qynJkW5Q==
                       integrity checksum failed when using sha512: wanted sha512-yXbYeFy+jUuYd3/CDcg2NkIYE991XYX/bje7LmjJigUciaeO1JR4XxXgCIV1/Zc/dRuFEyw1L0pbA+qynJkW5Q==
                       but got sha512-U1riIR+lBSNi3IbxtaHOIKdH8sLFv3NYfNv8sg7ZsNhcfl4HF2++BfqqrNAxoCLQW1iiylOj76ecnaUxz+z9yw==

Example failed deployment (or see any of the recent failures in #9067):
https://app.netlify.com/sites/apollo-client-docs/deploys/619d6dc222deda1694aa37a6

Although the error doesn't mention the package name, I was able to grep
for the expected hash in package-lock.json, which pointed to the `acorn`
package. As you can tell from this commit's changes, the actual hash
displayed above (ending in +z9yw==) happens to be the hash for
acorn@8.6.0, whereas the expected hash was that of acorn@8.5.0.