Change default Accept to application/json
While the GraphQL spec does not specify JSON as the serialization format, it it
by far the most popular. Therefore, it we can provide a better default accept
header, especially given we already default to "application/json" for the
content-type header. Consumers continue to have the ability to override this if
necessary.