Call cache.transformForLink before each link execution.
This is not a disruptive change for InMemoryCache, since it uses the
default implementation of tranformForLink (inherited from ApolloCache),
which simply returns the given document.

If you're using a different ApolloCache implementation that has a custom
transformForLink implementation, this new behavior should be more
convenient and flexible, but you should probably double-check that your
transformForLink method is suitably cached/idempotent, so multiple calls
with the same input document return the same (===) transformed document.