Update subscriptions.mdx
The parameters in the example is wrong according to the documentation of the SubscriptionClient in the subscriptions-transport-ws repo. 

From: https://github.com/apollographql/subscriptions-transport-ws

👉 Constructor(url, options, webSocketImpl) 👈

url: string : url that the client will connect to, starts with ws:// or wss://
options?: Object : optional, object to modify default client behavior
timeout?: number : how long the client should wait in ms for a keep-alive message from the server (default 30000 ms), this parameter is ignored if the server does not send keep-alive messages. This will also be used to calculate the max connection time per connect/reconnect
minTimeout?: number: the minimum amount of time the client should wait for a connection to be made (default 1000 ms)
lazy?: boolean : use to set lazy mode - connects only when first subscription created, and delay the socket initialization
connectionParams?: Object | Function | Promise<Object> : object that will be available as first argument of onConnect (in server side), if passed a function - it will call it and send the return value, if function returns as promise - it will wait until it resolves and send the resolved value.
reconnect?: boolean : automatic reconnect in case of connection error
reconnectionAttempts?: number : how much reconnect attempts
connectionCallback?: (error) => {} : optional, callback that called after the first init message, with the error (if there is one)
inactivityTimeout?: number : how long the client should wait in ms, when there are no active subscriptions, before disconnecting from the server. Set to 0 to disable this behavior. (default 0)
webSocketImpl?: Object - optional, constructor for W3C compliant WebSocket implementation. Use this when your environment does not have a built-in native WebSocket (for example, with NodeJS client)