Remove nagging updateQuery deprecation warning.
Now that fetchMore's updateQuery callback is implemented in terms of the
supported/documented cache.updateQuery method, I feel better about
allowing fetchMore to continue to take an updateQuery callback.

Also, everyone with any ability to migrate from updateQuery to
InMemoryCache field policies has presumably already done so, so this
warning is less useful now than it was following the release of AC3.