Move fetchBlockingPromise from WatchQueryOptions to reobserve parameter.
WatchQueryOptions.fetchBlockingPromise was introduced recently in #9599,
and probably needed more time to bake on the release-3.6 branch before
the final v3.6.0 release.

This commit removes fetchBlockingPromise from the public
WatchQueryOptions API, but continues passing an equivalent
Promise<boolean> when calling observable.reobserve (whenever the
useQuery options change).

This means we no longer block fetches for the initial call to useQuery,
but that should already be handled (even in <React.StrictMode>) by
waiting until the ObservableQuery gets its very first subscriber (in
useEffect) to initiate the first network request.

Since options.fetchBlockingPromise was intended for internal use and was
never documented, I am hopeful we still have time to remove it in a
patch version (likely 3.6.1).