Disable canUseDOM only when window.document comes from JSDOM.
I believe this is a more precise change than my previous commit, which
used isNode to determine canUseDOM. In most server-side rendering
setups, there will be no window.document.createElement, so canUseDOM
will already be false for that reason. It's only in Node.js environments
using jsdom that window.document.createElement might seem to be defined,
so that's the only case we really need to check.