Replace crypto-hash with a custom hashing function in tests
It looks like some combination of the changes in #9928 and #9942 exposed
a long running issue caused by using `crypto-hash` in our
`@apollo/client/link/persisted-queries` tests. `crypto-hash` spawns a
thread the first time it's called, and keeps that thread alive until the
end of program execution, if worker threads are available (which they are
for our tests since they're running in CI via Node). Unfortunately there
appears to be some kind of a race condition between `crypto-hash`
`unref`ing its worker and Jest reporting a successful test suite exit, which
leads to Jest reporting the following (which is seen by running jest with
the `--detectOpenHandles` setting enabled):

```
Jest has detected the following X open handles potentially keeping Jest from exiting:
```

This then appears to be leading to failed tests in CI.

We can look further into what's causing this race condition between
`crypto-hash` and Jest, but just dropping the use of `crypto-hash` is
likely a faster solution. We just need to verify that a hashing function can
be successfully set and used with the persisted queries link, which means we
can replace `crypto-hash` with our own simple function that doesn't use
worker threads. This commit does just that by introducing a super simple
custom `sha256` function that leverages Node `crypto`, and uses it
instead.

Fixes #9982