Experiment with not caching the Filesize Circle CI job.
Looking at a recent job, restoring and later saving the 400MB+ npm cache
takes 13 and 14 seconds, respectively:
https://app.circleci.com/pipelines/github/apollographql/apollo-client/17326/workflows/4db5bc1c-718e-4f0b-b6ec-7089b9aa8b55/jobs/94670

This job took a total of 1 minute 11 seconds, so restoring/saving the
cache accounts for ~38% of the total time.

If removing the caching altogether does not slow anything else down by
that much (27 total seconds), then the caching was not really delivering
on its intended goals, and should be removed.

Update after running CI on this commit:
https://app.circleci.com/pipelines/github/apollographql/apollo-client/17327/workflows/0c21e96a-736e-48f2-b2a9-55dea5405342/jobs/94672

Now only 1 minute 1 second (noticeably less time without any caching)!

The "Spin up environment" step went from taking 2s to 19s, because the
circleci/node@16.13.1 Docker image needed to be re-pulled. This seems
like something that could be addressed separately, if we care.

The `npm ci` step takes a little longer, 9s instead of 4s, but that's
presumably because everything has to be installed without a cache.
Still, this is much faster than restoring `node_modules` from cache.