Fix shouldRemoveField helper function in removeDirectivesFromDocument.
The hasRemoveDirective logic was broken because it did not enforce that
the _matching_ directive config has a `config.remove` property.

This led to the inappropriate removal of directives without
`config.remove`, in situations when any other configs are present with
`config.remove` set to `true`, even if those other configs do not match
any directives found on the current AST node.