Improve TypeScript type-safety of `cache.modify` (#10895)
* feat(cache): accept DELETE and INVALIDATE modifiers in the return type

Problem: The `DELETE` and `INVALIDATE` modifiers type is `any`.
Returning them from the `Modifier<T>` function triggers the
`@typescript-eslint/no-unsafe-return` ESLint rule [0],
since the modifier is then returning a value with type `any`.

Solution: Set the types of these modifiers to two different `unique
symbol`s. These serve as unique opaque types. The consuming code should
not peek into them. These 2 types are permitted in the return type of
the `Modifier<T>` function, since they are not `any` anymore, and thus,
are not assignable to `T`.

This change only affects TypeScript types. It has no runtime impact.

[0]: https://typescript-eslint.io/rules/no-unsafe-return/

* feat(cache): type inference for cache `Modifiers` type

Problem: The `fields` property in `cache.modify` did not offer ways to
infer the field names or values based on the field type. It accepted any
field names and the field values were `any`.

Solution: Add a generic type parameter for the object type to the
`Modifiers` type. The code can now use the `satisfies Modifiers<...>`
operator to inform TypeScript about the possible field names and values.

Field values include `Reference`s for objects and arrays. The consuming
code is then responsible for checking (or asserting) that the field
value is or is not a reference.

* refactor(cache): add default value for Modifiers generic parameter

Makes the change backwards-compatible.

* refactor(cache): use interfaces for invalidate and delete modifier types

Interfaces retain their names in error messages and hover dialogs. Type
aliases did not.

* refactor(cache): add generic parameter for ModifyOptions

A generic parameter is easier to discover and set than
`fields: { ... } satisfies Modifiers<...>`

* refactor(cache): use Record<string, any> as the default Entity type

Use a more appropriate real-life default type for `cache.modify`.

* code review adjustments

* refactor(cache-test): replace assertType with expect-type library

Use an already installed library instead of introducing a new utility
function.

* chore: use generic parameter instead of satisfies in changeset

Promote the use of a generic parameter of `cache.modify` over using
`satisfies` which requires importing the `Modifers` type.

* fix(cache): use AllFieldsModifier type in cache.modify parameter

---------

Co-authored-by: Lenz Weber-Tronic <lorenz.weber-tronic@apollographql.com>